blocklista

+++ 8JWT/app/resources/user.py
@@ -1,10 +1,11 @@
 from flask.views import MethodView
+from flask_jwt_extended import jwt_required
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 from db import db
 from passlib.hash import pbkdf2_sha256
-from flask_jwt_extended import create_access_token, create_refresh_token
-
+from flask_jwt_extended import create_access_token, create_refresh_token, get_jwt
+from blocklist import BLOCKLIST
 from models import UserModel
 from schemas import UserSchema
 
@@ -42,12 +43,20 @@ class UserLogin(MethodView):
 
 @blp.route("/user/<int:user_id>")
 class UserID(MethodView):
+    @jwt_required()
     @blp.response(200, UserSchema)
     def get(self, user_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required")
         return UserModel.query.get_or_404(user_id)
 
+    @jwt_required()
     @blp.response(200)
     def delete(self, user_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required")
         user = UserModel.query.get_or_404(user_id)
         try:
             db.session.delete(user)
@@ -55,3 +64,19 @@ class UserID(MethodView):
         except SQLAlchemyError as e:
             abort(500, message=e)
         return {"message": "User successfully deleted"}
+
+@blp.route("/logout")
+class LogoutUser(MethodView):
+    @jwt_required()
+    @blp.response(200)
+    def post(self):
+        jti = get_jwt()["jti"]
+        BLOCKLIST.add(jti)
+        return {"message": "Successfully logged out"}
+
+@blp.route("/refresh")
+class RefreshToken(MethodView):
+    @jwt_required()
+    @blp.response(200)
+    def post(self):
+        pass
